+++ jbpm-installer/db/ddl-scripts/sqlserver/sqlserver-jbpm-schema.sql
@@ -25,7 +25,7 @@
         priority int not null,
         processId varchar(255),
         processInstanceId numeric(19,0) not null,
-        processSessionId int not null,
+        processSessionId numeric(19,0) not null,
         status varchar(255),
         taskId numeric(19,0),
         primary key (id)
@@ -63,7 +63,7 @@
     create table ContextMappingInfo (
         mappingId numeric(19,0) identity not null,
         CONTEXT_ID varchar(255) not null,
-        KSESSION_ID int not null,
+        KSESSION_ID numeric(19,0) not null,
         OWNER_ID varchar(255),
         OPTLOCK int,
         primary key (mappingId)
@@ -278,7 +278,7 @@
     );
 
     create table SessionInfo (
-        id int identity not null,
+        id numeric(19,0) identity not null,
         lastModificationDate datetime,
         rulesByteArray image,
         startDate datetime,
@@ -314,7 +314,7 @@
         previousStatus int,
         processId varchar(255),
         processInstanceId numeric(19,0) not null,
-        processSessionId int not null,
+        processSessionId numeric(19,0) not null,
         skipable bit not null,
         status varchar(255),
         workItemId numeric(19,0) not null,
