+++ jbpm-installer/db/ddl-scripts/sqlserver2008/sqlserver2008-jbpm-schema.sql
@@ -25,7 +25,7 @@
         priority int not null,
         processId varchar(255),
         processInstanceId bigint not null,
-        processSessionId int not null,
+        processSessionId bigint not null,
         status varchar(255),
         taskId bigint,
         primary key (id)
@@ -63,7 +63,7 @@
     create table ContextMappingInfo (
         mappingId bigint identity not null,
         CONTEXT_ID varchar(255) not null,
-        KSESSION_ID int not null,
+        KSESSION_ID bigint not null,
         OWNER_ID varchar(255),
         OPTLOCK int,
         primary key (mappingId)
@@ -278,7 +278,7 @@
     );
 
     create table SessionInfo (
-        id int identity not null,
+        id bigint identity not null,
         lastModificationDate datetime2,
         rulesByteArray varbinary(MAX),
         startDate datetime2,
@@ -314,7 +314,7 @@
         previousStatus int,
         processId varchar(255),
         processInstanceId bigint not null,
-        processSessionId int not null,
+        processSessionId bigint not null,
         skipable bit not null,
         status varchar(255),
         workItemId bigint not null,
